requirements.txt
torch >= 1.9.0
torchaudio >= 0.9.0
torchvision >= 0.10.0
torchmetrics <= 0.7.3
pytorch-lightning >= 1.4.0, < 1.5.0
lightning-bolts >= 0.3.3
pillow >= 8.1
pysoundfile >= 0.9.0; platform_system == "Windows"
numpy >= 1.20.1
einops >= 0.3.0
av >= 8.0.3
pandas >= 1.3.1
ipython >= 7.26.0
pygments >= 2.9.0
matplotlib >= 3.3.0
tqdm >= 4.49.0
protobuf <= 3.19.0
ffmpeg-python >= 0.2.0
scipy>=1.0.0